benchmarks/rag-framework/requirements.txt
# ── Core utilities ────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
numpy>=1.26.0
rich>=13.7.0
tqdm>=4.66.0

# ── LLM (OpenRouter via OpenAI-compatible client) ─────────────────────────────
openai>=1.30.0

# ── Embeddings (local, no API key needed) ─────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.0.0           # backend for sentence-transformers

# ── RAG Framework: LlamaIndex ─────────────────────────────────────────────────
llama-index-core>=0.11.0
llama-index-llms-openai>=0.3.0
llama-index-embeddings-huggingface>=0.3.0

# ── RAG Framework: LangChain ──────────────────────────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0       # in-memory vector store for LangChain PoC

# ── RAG Framework: Haystack ───────────────────────────────────────────────────
haystack-ai>=2.5.0
